Api code arrange
