+++ imports/modules/flows/server/api.js
@@ -1,32 +1,14 @@
 import { Router } from 'meteor/iron:router'
-import { authenticate } from '../../_common/server/api'
+import { authenticate, catchHandler, reply } from '../../_common/server/api'
 import { Flows } from '../both/collection'
-
-/**
- * 
- * @param {Object} res 
- * @param {Object} data 
- * @param {Number} status 
- */
-const reply = (res, data, status) => {
-  if (typeof data === 'object') {
-    res.end(JSON.stringify(data, null, 2))
-  }
-  else {
-    res.writeHead(status || 200)
-    res.end(data)
-  }
-}
 
 Router.route('/api/flows', function () {
   const req = this.request
   const res = this.response
 
   authenticate(req.headers)
-    .then(u => {
-      let flowsQuery = {
-        user: u._id
-      }
+    .then(user => {
+      let flowsQuery = { user: user._id }
 
       if (req.query.triggerType) {
         flowsQuery['trigger.type'] = req.query.triggerType
@@ -38,9 +20,6 @@ Router.route('/api/flows', function () {
 
       reply(res, flows)
     })
-    .catch(ex => {
-      console.error(ex)
-      reply(res, ex.reason, 314)
-    })
+    .catch(ex => catchHandler(res, ex))
 
 }, {where: 'server'})
